WPF
In order to analyze the geothermal potential in Hangjiahu Plain, this paper tentatively proposes to use analytic hierarchy process to construct the geothermal evaluation index system in the study area, collect, process and generate several factor maps, which are essential data for generating AHP geothermal potential map. We use these factor graphs to compile a C# WPF form program for user interaction according to the corresponding weight coefficients given by experts. Based on the .NET framework, we call the GDAL library to weight and synthesize the factor graph to generate a geothermal potential graph.